Update WASM setup
- Update docs/READMEs
- Remove any mention of ASM.js as it is now only needed for very old
  browser and it is not something we are preoccupied with now.

Close #90
