+++ .github/workflows/build.yml
@@ -149,11 +149,11 @@ jobs:
     - name: Test C FFI
       run: make -C lib ../target/test/c.stamp
 
-    - name: Test Node.js package
+    - name: Test Node.js (Neon) package
       run: make -C lib ../target/test/node.stamp
 
     - name: Test WASM package
-      run: make -C lib ../target/test/wasm.stamp 
+      run: make -C lib ../target/test/wasm.stamp
 
     - name: Test JNI
       run: make -C lib ../target/test/java.stamp
