+++ lib/web/README.md
@@ -1,14 +1,70 @@
+[![](https://img.shields.io/npm/v/@spruceid/didkit-wasm?label=%40spruceid%2Fdidkit-wasm&logo=npm)](https://www.npmjs.com/package/@spruceid/didkit-wasm) [![](https://img.shields.io/npm/v/@spruceid/didkit-wasm-node?label=%40spruceid%2Fdidkit-wasm-node&logo=npm)](https://www.npmjs.com/package/@spruceid/didkit-wasm-node)
+<!-- Might want those badge in the main README. -->
+
 # DIDKit WASM
+
+## Prerequisites to Build from Source
+
+NPM packages are available but if you would like to compile DIDKit yourself
+(e.g. to enable different cryptographic backends) you will need the WASM
+compiler toolchain as well as a specific build tool:
+
+```bash
+$ rustup target add wasm32-unknown-unknown
+$ cargo install wasm-pack
+# OR
+# $ curl https://rustwasm.github.io/wasm-pack/installer/init.sh -sSf | sh
+```
+
+## Installation and Usage
+
+### Node
+
+WASM can be used in Node.js (any recent version):
+```bash
+$ npm i @spruceid/didkit-wasm-node
+```
+
+Or build it from source:
+```bash
+$ wasm-pack build --target nodejs
+```
+
+### Web Frameworks (Bundled)
+
+WASM can be used with web frameworks and bundlers like Webpack:
+```bash
+$ npm i @spruceid/didkit-wasm
+```
+
+Or build it from source:
+```bash
+$ wasm-pack build
+```
+
+> If Webpack doesn't work with the default configuration, you can have a look at
+> our configuration for
+> [tzprofiles](https://github.com/spruceid/tzprofiles/blob/main/dapp/webpack.config.js).
+
+### Vanilla Javascript
+
+WASM can be used with plain Javascript with newer browsers. As it cannot be used
+as a NPM package you have to build it manually:
+```bash
+$ wasm-pack build --target web
+```
+
+The manual tests in `test/` serve as an example on how to import DIDKit.
+
+## Tests
+
+The `test/` directory contains manual tests to run in the browser. Instructions
+are in the README of the directory.
+
+## Non-Default Compilation
 
 _**The current version of the `ring` crate does not provide all the symbols
 needed to run on the browser, see DEPS.md**_
-
-
-To install `wasm-pack` run:
-
-```bash
-cargo install wasm-pack
-```
 
 To compile all features plus `wasm32_c` on `ring`, a C compiler is needed, see
 [spruceid/ssi](https://github.com/spruceid/didkit/tree/wasm):
@@ -20,47 +76,14 @@ sudo apt install clang-10 llvm-10
 
 Then to compile with all features:
 ```bash
-TARGET_CC=clang-10 TARGET_AR=llvm-ar-10 wasm-pack build --target web --out-dir pkg/wasm
+TARGET_CC=clang-10 TARGET_AR=llvm-ar-10 wasm-pack build --out-dir pkg
 ```
 
 To use a custom subset of features:
 ```bash
-wasm-pack build --target web --out-dir pkg/wasm -- --no-default-features --features=issue        # issue credential/presentation
-wasm-pack build --target web --out-dir pkg/wasm -- --no-default-features --features=verify       # verify credential/presentation
-wasm-pack build --target web --out-dir pkg/wasm -- --no-default-features --features=credential   # issue/verify credential
-wasm-pack build --target web --out-dir pkg/wasm -- --no-default-features --features=presentation # issue/verify presentation
+wasm-pack build --out-dir pkg -- --no-default-features --features=issue        # issue credential/presentation
+wasm-pack build --out-dir pkg -- --no-default-features --features=verify       # verify credential/presentation
+wasm-pack build --out-dir pkg -- --no-default-features --features=credential   # issue/verify credential
+wasm-pack build --out-dir pkg -- --no-default-features --features=presentation # issue/verify presentation
 ```
 *don't forget to add `TARGET_CC` and `TARGET_AR` if using `ring` with `wasm32_c`*
-
-
-# ASM target
-For the ASM target the `wasm2js` build is necessary, to get this tool
-follow the steps in [binaryen](https://github.com/WebAssembly/binaryen#building).
-
-To compile all features for WASM target:
-```bash
-wasm-pack build --target web --out-dir pkg/wasm
-npm --prefix loader/wasm install
-npm --prefix loader/wasm run build
-```
-
-To compile all features for ASM target, assuming you're in this folder:
-```bash
-wasm-pack build --target bundler --out-dir pkg/asmjs
-cd pkg/asmjs
-$(BINARYEN_ROOT)/bin/wasm2js --pedantic -o didkit_wasm_bg1.js didkit_wasm_bg.wasm
-cd ../..
-cd loader/asmjs
-./repack.sh
-npm install
-npm run build
-```
-
-# WASM tests
-
-`loader/wasm` contains tests. To run them you need to serve the files and open
-the page in a browser (you can then see the results in the console). You can
-start a server with
-```bash
-$ npm run serve
-```
