--- lib/web/loader/asmjs/.gitignore
@@ -1,3 +1,0 @@
-node_modules
-didkit-output.js
-didkit-asm.min.js
