--- lib/web/loader/asmjs/didkit_wasm_bg1.js
@@ -1,1 +1,0 @@
-../../pkg/asmjs/didkit_wasm_bg1.js
