--- lib/web/loader/asmjs/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "didkit-asm-loader",
-  "author": "Spruce Systems, Inc.",
-  "collaborators": [
-    "Tiago Nascimento <tiago.nascimento@spruceid.com>",
-    "Gregorio Granado Magalhaes <gregorio.magalhaes@spruceid.com"
-  ],
-  "version": "0.1.0",
-  "files": [
-    "asm-loader.js"
-  ],
-  "module": "asm-loader.js",
-  "sideEffects": false,
-  "scripts": {
-    "build": "webpack"
-  },
-  "devDependencies": {
-    "webpack": "^5.18.0",
-    "webpack-cli": "^4.4.0"
-  }
-}
