--- lib/web/loader/wasm/.eslintrc.js
@@ -1,14 +1,0 @@
-module.exports = {
-  env: {
-    browser: true,
-    es2021: true,
-  },
-  parserOptions: {
-    ecmaVersion: 12,
-    sourceType: "module",
-  },
-  plugins: ["prettier"],
-  rules: {
-    "prettier/prettier": "error",
-  },
-};
