--- lib/web/loader/wasm/.gitignore
@@ -1,4 +1,0 @@
-node_modules
-didkit-loader.min.js
-didkit-loader.min.js.LICENSE.txt
-didkit-loader.min.js.map
