--- lib/web/loader/wasm/didkit_wasm.js
@@ -1,1 +1,0 @@
-../../pkg/wasm/didkit_wasm.js
