--- lib/web/loader/wasm/didkit_wasm_bg.wasm
@@ -1,1 +1,0 @@
-../../pkg/wasm/didkit_wasm_bg.wasm
