--- lib/web/loader/wasm/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "didkit-wasm-loader",
-  "author": "Spruce Systems, Inc.",
-  "collaborators": [
-    "Tiago Nascimento <tiago.nascimento@spruceid.com>"
-  ],
-  "version": "0.1.0",
-  "files": [
-    "didkit-loader.js"
-  ],
-  "module": "didkit-loader.js",
-  "sideEffects": false,
-  "scripts": {
-    "build": "webpack --mode development",
-    "serve": "ws"
-  },
-  "devDependencies": {
-    "eslint": "^7.22.0",
-    "eslint-config-prettier": "^7.2.0",
-    "eslint-config-standard": "^16.0.2",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prettier": "^3.3.1",
-    "eslint-plugin-promise": "^4.3.1",
-    "prettier": "2.2.1",
-    "webpack": "^5.26.2",
-    "webpack-cli": "^4.5.0",
-    "local-web-server": "^4.2.1"
-  }
-}
