+++ lib/web/test/index.html
@@ -2,8 +2,7 @@
   <head>
     <title>DIDKit WASM Example</title>
     <meta charset="utf-8">
-    <script src="/didkit-loader.min.js"></script>
-    <script src="/test/index.test.js"></script>
+    <script type="module" src="index.test.js"></script>
   </head>
 
   <body>
