+++ lib/web/test/package.json
@@ -1,0 +1,1 @@
+{}
