+++ lib/web/test/pkg
@@ -1,0 +1,1 @@
+../pkg
